Create PerfectShuffle(C++)
I saw online that if you perfectly shuffled a deck of cards 8 times, the cards would return to their original order. This intrigued me, so I tried doing this with a deck of cards. After 20 minutes trying to perfectly shuffle a deck of cards 8 times, I realized I had wasted my time. So I decided to make a program that perfectly shuffles a deck of cards every time without fail. 

In this program you can choose how many times that you want to perfectly shuffle the deck of cards. Turns out that perfectly shuffling a deck of cards 8 times actually does return the cards to the original order.