Create InvestmentAnalysisP&L(VBA)
I took an Investment Analysis course this year in school. Throughout the year, we paper traded stocks as a project for the class. My professor wanted us to create a trade log and submit it to him at the end of the year. Most students just took screenshots of their trades and sent it to our professor.

I created an excel sheet with my trades, account balance, and the S&P500 closing price for each day. I also had a section of my workbook dedicated to my Profit/Loss on each ticker that I traded. Once I put in new trades into my log, I could hit a button that would trigger a macro and it would automatically update my P&L on each ticker. The workbook was the conditionally formatted to show the profits in green and the losses in red.