Create TicTacToe Vs Computer (C++)
Went to the mountains this weekend with my family. We had a rainy day and were watching movies. I decided to make a Tic Tac Toe game since I was bored. The computer randomly puts down a place on the board. I may revisit this to give the computer some strategy to make this game a little tougher.