Create PasswordVerifier
I'm starting to look at things as if I were software engineer while I'm on the internet. When creating an account for a website today I was typing in my password and thought, I could do this. I had some time to burn tonight and decided to challenge myself to see if I really could do it. So, I made a program that makes sure you type in a password with the right parameters and then confirms that you know what you typed in.